requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
facebook-sdk==2.0.0
mysql-connector==2.1.4
numpy==1.12.0
packaging==16.8
pandas==0.19.2
pkg-resources==0.0.0
pymssql==2.1.3
pyparsing==2.2.0
python-dateutil==2.6.0
python-nmap==0.6.1
pytz==2016.10
requests==2.13.0
selenium==3.3.1
six==1.10.0
vk-requests==0.9.7
